pyproject.toml
[tool.poetry]
name = "discord-bot-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
dnspython = "^2.2.1"
pymongo = "^4.0.2"
python-dotenv = "^0.19.2"
Flask = "^2.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
